Articles for creators
Add number of articles to creators page, and also count articles in
the first and last issue a creator has contributed to.

Sice the data query for that initially took ~ 20 s to execute,
optimize the data in two ways: First add a column "magic" to the
issues table for storing the previously calculated single integer
issue descriptor, and then add a materialized view for the entire
query.

Use the precalculated "magic" value to optimize some queries
throughout the application.