Improve server filter structure.
First match the url, then the method (so errors are 404 and not bad
method), and only then clone the db pool (to avoid unnessesary work,
even if it's only an Arc::clone).

This implies that the `PgPool` is now the last argument to handlers,
rather than the first.

Also, postrelease.